refactor: rename IconButtons to just IconButton
