+++ dev-client/src/components/SiteCard.tsx
@@ -20,7 +20,7 @@ import {StyleSheet, ViewStyle} from 'react-native';
 
 import {Site} from 'terraso-client-shared/site/siteSlice';
 
-import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButtons';
+import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButton';
 import {Card} from 'terraso-mobile-client/components/Card';
 import {
   Box,
