+++ dev-client/src/components/buttons/icons/BadgedIconButton.tsx
@@ -18,7 +18,7 @@ import {PressableProps} from 'react-native';
 
 import {Pressable} from 'native-base';
 
-import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButtons';
+import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButton';
 import {IconName} from 'terraso-mobile-client/components/icons/Icon';
 import {Badge} from 'terraso-mobile-client/components/NativeBaseAdapters';
