+++ dev-client/src/components/buttons/icons/CloseButton.tsx
@@ -21,7 +21,7 @@ import {PressableProps, View} from 'react-native';
 import {
   IconButton,
   IconButtonVariant,
-} from 'terraso-mobile-client/components/buttons/icons/IconButtons';
+} from 'terraso-mobile-client/components/buttons/icons/IconButton';
 
 type CloseButtonProps = {
   variant?: IconButtonVariant;
