+++ dev-client/src/components/buttons/icons/SettingsButton.tsx
@@ -18,7 +18,7 @@
 import {forwardRef} from 'react';
 import {PressableProps, View} from 'react-native';
 
-import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButtons';
+import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButton';
 
 type BackButtonProps = {
   onPress?: PressableProps['onPress'];
