+++ dev-client/src/screens/ColorAnalysisScreen/ColorAnalysisHomeScreen.tsx
@@ -23,7 +23,7 @@ import {Fab} from 'native-base';
 
 import {updateDepthDependentSoilData} from 'terraso-client-shared/soilId/soilIdSlice';
 
-import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButtons';
+import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButton';
 import {Icon} from 'terraso-mobile-client/components/icons/Icon';
 import {
   decodeBase64Jpg,
