+++ dev-client/src/screens/ProjectInputScreen/DepthTable.tsx
@@ -24,7 +24,7 @@ import {
   ProjectDepthInterval,
 } from 'terraso-client-shared/soilId/soilIdSlice';
 
-import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButtons';
+import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButton';
 import {ConfirmDeleteDepthModal} from 'terraso-mobile-client/components/modals/ConfirmDeleteDepthModal';
 import {Box} from 'terraso-mobile-client/components/NativeBaseAdapters';
 import {DataGridTable} from 'terraso-mobile-client/components/tables/DataGridTable';
